Prevent programs from changing vertical resolution
This commit intercepts all attempts to change vertical resolution via
int 10h, pretending to succeed every time but actually doing nothing.

Doing this for *every* attempt to change vertical resolution is probably
not the best solution long term. For example, a palette-only TSR ought
not to touch anything related to screen geometry at all. But this is a
start, and it allows the use of 16px fonts in DOS programs that try to
set a 14px EGA mode.